requirements.txt
argcomplete==1.11.1
asgiref==3.2.7
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
cfn-flip==1.2.2
chardet==3.0.4
click==7.1.1
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
docutils==0.15.2
durationpy==0.5
future==0.18.2
hjson==3.0.1
idna==2.9
importlib-metadata==1.6.0
jmespath==0.9.5
kappa==0.6.0
Markdown==3.2.1
mysqlclient==1.4.6
pip-tools==4.5.1
placebo==0.9.0
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.0
tqdm==4.45.0
troposphere==2.6.0
urllib3==1.25.8
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
